Wrap up SAFT to provide default startup environment.
This CL does the following:

- introduce a shell script which can be used to start
  SAFT from a standard released image

- add code to retrieve the root public key from the
  firmware running at the beginning of the test (needed
  since the keys are not available in /etc/devkeys anymore)

- add a test case to test key retrieval code


BUG=chromium-os:4801
TEST=see below:

- install the new generated ChromeOS image on a ZGA
- leave the flash drive plugged in
- restart the machine
- once it comes up place a new firmware image into
  /tmp/ZGA_010b.fd
- invoke SAFT as follows:
  sudo /usr/sbin/firmware/saft/runtests.sh /tmp/ZGA_010b.fd
- wait til the test finishes running

Expected results: right after the start the untitest success
reports are printed on the console. Then the machine resent several times. Once it settles, the following is found in /media/C-STATE/var/saft_log.txt:
localhost ~ # cat /media/C-STATE/var/saft_log.txt
11:27:26 AM: Automated firmware test log generated on Aug 13 2010
11:27:26 AM: Executing rootdev
11:27:26 AM: Original boot state 1:1:0:0:3
11:27:30 AM: Executing vbutil_firmware --verify /tmp/tmpOMbIY3/var/.fw_test/VBOOTA --signpubkey /tmp/tmpOMbIY3/var/.fw_test/root.pubkey  --fv /tmp/tmpOMbIY3/var/.fw_test/FVMAIN
11:27:31 AM: Executing vbutil_firmware --verify /tmp/tmpOMbIY3/var/.fw_test/VBOOTB --signpubkey /tmp/tmpOMbIY3/var/.fw_test/root.pubkey  --fv /tmp/tmpOMbIY3/var/.fw_test/FVMAINB
11:27:31 AM: Executing vbutil_firmware --verify /tmp/tmpOMbIY3/var/.fw_test/VBOOTA --signpubkey /tmp/tmpOMbIY3/var/.fw_test/root.pubkey  --fv /tmp/tmpOMbIY3/var/.fw_test/FVMAIN
11:27:31 AM: Executing vbutil_firmware --verify /tmp/tmpOMbIY3/var/.fw_test/VBOOTB --signpubkey /tmp/tmpOMbIY3/var/.fw_test/root.pubkey  --fv /tmp/tmpOMbIY3/var/.fw_test/FVMAINB
11:27:31 AM: Executing df /tmp/tmp.XhhtrzKqdA/usr/sbin/firmware/saft
11:27:31 AM: Executing blkid
11:27:31 AM: Executing rootdev
11:27:31 AM: Executing mount
11:27:31 AM: Handling //etc/init/fw_test.conf
11:27:31 AM: Executing rootdev
11:27:31 AM: Executing mount
11:27:31 AM: Handling /tmp/tmpF5QjlD/etc/init/fw_test.conf
11:27:31 AM: Executing rootdev
11:27:31 AM: Executing mount
11:27:31 AM: Handling /tmp/tmp.XhhtrzKqdA/etc/init/fw_test.conf
11:27:31 AM: program the new image
11:28:02 AM: restart
11:28:02 AM: Executing reboot
11:28:28 AM: Executing cgpt show /dev/sda
11:28:28 AM: Executing rootdev
11:28:28 AM: Rebooted into state 1:1:0:0:3 on step 0
11:28:28 AM: calling <bound method FirmwareTest.set_try_fw_b of <__main__.FirmwareTest object at 0x6f777dcc>>
11:28:28 AM: Requesting restart with FW B
11:28:28 AM: Executing reboot_mode --try_firmware_b=1
11:28:28 AM: Executing reboot
11:28:45 AM: Executing cgpt show /dev/sda
11:28:45 AM: Executing rootdev
11:28:45 AM: Rebooted into state 1:2:0:0:3 on step 1
11:28:45 AM: Executing reboot
11:29:02 AM: Executing cgpt show /dev/sda
11:29:02 AM: Executing rootdev
11:29:02 AM: Rebooted into state 1:1:0:0:3 on step 2
11:29:02 AM: calling <bound method FirmwareTest.corrupt_firmware of <__main__.FirmwareTest object at 0x6f6dedcc>> with parameter a
11:29:02 AM: Corrupting firmware a
11:29:44 AM: Executing reboot
11:30:02 AM: Executing cgpt show /dev/sda
11:30:02 AM: Executing rootdev
11:30:02 AM: Rebooted into state 1:2:0:0:3 on step 3
11:30:02 AM: calling <bound method FirmwareTest.restore_firmware of <__main__.FirmwareTest object at 0x6f732dcc>> with parameter a
11:30:02 AM: Restoring firmware a
11:30:44 AM: Executing reboot
11:31:00 AM: Executing cgpt show /dev/sda
11:31:00 AM: Executing rootdev
11:31:00 AM: Rebooted into state 1:1:0:0:3 on step 4
11:31:00 AM: calling <bound method FirmwareTest.corrupt_kernel of <__main__.FirmwareTest object at 0x6f6abdcc>> with parameter a
11:31:00 AM: Corrupting kernel a
11:31:00 AM: Executing dd if=/dev/sda2 of=/tmp/tmpqq9dif/var/.fw_test/kernel_header_dump bs=1 count=1
11:31:00 AM: Executing dd if=/tmp/tmpqq9dif/var/.fw_test/kernel_header_dump of=/dev/sda2 bs=1 count=1
11:31:00 AM: Executing reboot
11:31:17 AM: Executing cgpt show /dev/sda
11:31:17 AM: Executing rootdev
11:31:17 AM: Rebooted into state 1:1:0:0:5 on step 5
11:31:17 AM: calling <bound method FirmwareTest.restore_kernel of <__main__.FirmwareTest object at 0x6f7eddcc>> with parameter a
11:31:17 AM: restoring kernel a
11:31:17 AM: Executing dd if=/dev/sda2 of=/tmp/tmp5EgX7C/var/.fw_test/kernel_header_dump bs=1 count=1
11:31:17 AM: Executing dd if=/tmp/tmp5EgX7C/var/.fw_test/kernel_header_dump of=/dev/sda2 bs=1 count=1
11:31:17 AM: Executing reboot
11:31:35 AM: Executing cgpt show /dev/sda
11:31:35 AM: Executing rootdev
11:31:35 AM: Rebooted into state 1:1:0:0:3 on step 6
11:31:35 AM: calling <bound method FirmwareTest.revert_firmware of <__main__.FirmwareTest object at 0x6f658dcc>>
11:31:35 AM: restoring original firmware image
11:31:35 AM: Executing flashrom -w /tmp/tmpCDuPOS/var/.fw_test/flashrom.bak
11:32:08 AM: Executing reboot
11:32:26 AM: Executing cgpt show /dev/sda
11:32:26 AM: Executing rootdev
11:32:26 AM: Rebooted into state 1:1:0:0:3 on step 7
11:32:26 AM: Removing upstart scripts
11:32:26 AM: Executing df /tmp/sdb3/usr/sbin/firmware/saft
11:32:26 AM: Executing blkid
11:32:26 AM: Executing rootdev
11:32:26 AM: Executing mount
11:32:26 AM: Executing mount -o remount,rw /dev/root
11:32:26 AM: Handling //etc/init/fw_test.conf
11:32:26 AM: Executing rootdev
11:32:26 AM: Executing mount
11:32:26 AM: Executing mount /dev/sda5 /tmp/tmpQqJq_X
11:32:26 AM: Handling /tmp/tmpQqJq_X/etc/init/fw_test.conf
11:32:26 AM: Executing umount /tmp/tmpQqJq_X
11:32:28 AM: Executing rootdev
11:32:28 AM: Executing mount
11:32:28 AM: Handling /tmp/sdb3/etc/init/fw_test.conf
11:32:28 AM: we are done!

Review URL: http://codereview.chromium.org/3128006

+++ saft/saft_utility.py
@@ -311,16 +311,7 @@ class FirmwareTest(object):
                 % expected_vector)
             sys.exit(1)
         if this_step == (len(self.test_state_sequence) - 1):
-            # We are almost there, no more steps to take.
-            if not self._verify_fw_id(FWID_BACKUP_FILE):
-                self.chros_if.log(
-                    'Error: Failed to restore to original firmware')
-                sys.exit(1)
-            self.chros_if.log('Removing upstart scripts')
-            self._handle_upstart_script(False)
-            self.chros_if.log('we are done!')
-            sys.exit(0)
-
+            self.finish_saft()
         if action and not self._verify_fw_id(FWID_NEW_FILE):
             self.chros_if.log('Error: Wrong FWID value')
             sys.exit(1)
@@ -334,6 +325,28 @@ class FirmwareTest(object):
                 self.chros_if.log('calling %s' % str(action))
                 action()
         self._set_step(this_step + 1)
+        self.chros_if.run_shell_command('reboot')
+
+    def finish_saft(self):
+        '''SAFT is done, restore the environment and expose the results.
+
+        Restore to the original firmware, delete the startup scripts from all
+        drives, copy the log to the stateful partition on the flash drive and
+        restart the machine.
+        '''
+        if not self._verify_fw_id(FWID_BACKUP_FILE):
+            self.chros_if.log(
+                'Error: Failed to restore to original firmware')
+            sys.exit(1)
+        self.chros_if.log('Removing upstart scripts')
+        self._handle_upstart_script(False)
+        self.chros_if.log('we are done!')
+
+        # Shut_down will delete the SAFT state directory, let's have the log
+        # copied one level up.
+        dummy = self.chros_if.state_dir_file('dummy')
+        log_dst_dir = os.path.realpath(os.path.join(os.path.dirname(dummy), '..'))
+        self.chros_if.shut_down(os.path.join(log_dst_dir, LOG_FILE))
         self.chros_if.run_shell_command('reboot')
 
 
@@ -466,9 +479,8 @@ def main(argv):
     if missing_execs:
         usage('Program(s) %s not found in PATH' % ' '.join(missing_execs), 1)
 
-    for req_param in 'image_file pub_key'.split(' '):
-        if req_param not in opt_dictionary:
-            usage(rv=1)
+    if 'image_file' not in opt_dictionary:
+        usage(retv=1)
 
     FST.init_fw_test(opt_dictionary, CHROS_IF)
     CHROS_IF.log('program the new image')
